chore: Add Markdown link checker configuration and workflow (#34)
# Pull Request

## Description

This change introduces a new GitHub Actions workflow step to check
Markdown links in the repository. The workflow uses the
UmbrellaDocs/action-linkspector action to validate links in Markdown
files.

A new configuration file `.github/other-configurations/.linkspector.yml`
has been added to specify the directories to scan, files to exclude, and
acceptable HTTP status codes for link validation.

The code-quality.yml workflow file has been updated to include this new
step, which runs on Ubuntu and uses a GitHub token for authentication.
The step is configured to fail on errors and provide feedback through
GitHub PR reviews.

This enhancement will help maintain the integrity of documentation links
throughout the project.

fixes #12