ci: Separate CodeLimit Job (#106)
# Pull Request

## Description

This pull request introduces a new job to the GitHub Actions workflow to
enforce code size limits. The most important changes include the
addition of a new job named `run-code-limit` in the
`.github/workflows/code-checks.yml` file.

Changes to GitHub Actions workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R100-R107):
Added a new job `run-code-limit` to the workflow, which includes steps
to checkout the repository and run the `getcodelimit/codelimit-action`.

Fixes #107
Fixes #108