ci: Add Zizmor Job (#116)
# Pull Request

## Description

This pull request includes the addition of a new job to the GitHub
Actions workflow in the `.github/workflows/code-checks.yml` file. The
new job is responsible for running the `zizmor` tool to check GitHub
Actions.

Enhancements to GitHub Actions workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R113-R137):
Added a new job `run-zizmor` to check GitHub Actions using the `zizmor`
tool. This job includes steps for checking out the repository,
installing the latest version of `uv`, running `zizmor`, and uploading
the SARIF file with the results.

Fixes #115