chore: Add Code of Conduct (#120)
# Pull Request

## Description

This pull request introduces a new Code of Conduct document to the
project. The document outlines the community's commitment to providing a
harassment-free experience for everyone and details the standards for
acceptable behavior, enforcement responsibilities, and the scope of the
Code of Conduct.

Key changes include:

* Addition of the Contributor Covenant Code of Conduct to the
`docs/CODE_OF_CONDUCT.md` file, which includes sections on pledges,
standards, enforcement responsibilities, scope, enforcement guidelines,
and attribution.

Fixes #121