ci: Update Dependency Review (#192)
# Pull Request

## Description

This pull request includes updates to the GitHub Actions workflow
configuration to enhance the dependency review process. The most
important changes are:

Enhancements to dependency review:

*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR49-R50):
Added write permissions for pull requests to the `dependency-review`
job.
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR59-R60):
Configured the `dependency-review` action to comment a summary in the
pull request on failure.