chore: Refactor Git Hooks (#195)
# Pull Request

## Description

This pull request focuses on improving the pre-commit hook checks by
consolidating the scripts and adding a new gitleaks check. The most
important changes include the addition of a gitleaks check in the
`Justfile`, the refactoring of the pre-commit script to run multiple
checks, and the removal of individual pre-commit scripts.

Enhancements to pre-commit hook checks:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR32-R38):
Added a new `gitleaks-detect` target to run gitleaks detection.
*
[`githooks/pre-commit`](diffhunk://#diff-d5fe9d324948511c959e14d0aa311bb057ce7236626bb86fa51351dc26845a4dL4-R30):
Refactored the script to run gitleaks, prettier, and format checks using
the `just` command, and removed the loop that executed individual
scripts.

Removal of individual pre-commit scripts:

*
[`githooks/pre-commit-scripts/gitleaks.sh`](diffhunk://#diff-b488692093f21df73d5c8d67f3a6df3273b7199a34ca84e9c9e1a115b20df271L1-L4):
Removed the script as its functionality is now included in the main
pre-commit script.
*
[`githooks/pre-commit-scripts/justfile.sh`](diffhunk://#diff-fcd01eccc95d37bac6ab073a69ec60e574b754774360a13f31e1a39466668fc9L1-L4):
Removed the script as its functionality is now included in the main
pre-commit script.
*
[`githooks/pre-commit-scripts/prettier.sh`](diffhunk://#diff-1e2ab1eef74462625d345d9811c6cd8e51996878583f4a9878ad8d62ec7d014dL1-L4):
Removed the script as its functionality is now included in the main
pre-commit script.

Fixes #194