chore: Remove CodeQL Configuration (#225)
# Pull Request

## Description

This pull request updates the CodeQL configuration by removing an unused
query filter and simplifying the workflow configuration. The most
important changes are as follows:

### CodeQL Configuration Updates:

*
[`.github/other-configurations/codeql-config.yml`](diffhunk://#diff-687a0af9316072cbdab79cdb046aac181231f8b5f3b5502bc796a9d949ce65beL1-L3):
Removed the `actions/unpinned-tag` query filter from the `query-filters`
section, as it is no longer needed.

### Workflow Simplification:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L110):
Removed the reference to the `codeql-config.yml` file in the
`config-file` parameter of the CodeQL analysis job, simplifying the
workflow setup.