chore: Update Dependabot Schedule (#236)
# Pull Request

## Description

This pull request updates the Dependabot configuration to switch from a
daily schedule to a cron-based schedule for dependency updates.

Changes to Dependabot scheduling:

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L10-R11):
Updated the `github-actions` group to use a cron schedule (`30 7 * * *`)
instead of the previous daily schedule.
*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L27-R27):
Updated the `typescript` group to use a cron schedule (`30 7 * * *`)
instead of the previous daily schedule.