chore: Add pinact commands (#252)
# Pull Request

## Description

This pull request introduces new commands to the `Justfile` for managing
`pinact` operations, alongside existing `zizmor-check` and Git hooks
commands.

New `pinact` commands added:

*
[`pinact-run`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR51-R66):
A command to execute `pinact run`.
*
[`pinact-check`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR51-R66):
A command to run `pinact` with `--verify` and `--check` options for
validation purposes.
*
[`pinact-update`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR51-R66):
A command to run `pinact` with the `--update` option for updating
configurations or dependencies.