ci: Add Lefthook Validate Job (#257)
# Pull Request

## Description

This pull request adds a new job to the GitHub Actions workflow to
validate Lefthook configurations. This ensures that the Lefthook setup
is correctly configured and adheres to the expected standards.

### Workflow Enhancements:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R91-R104):
Added a new `lefthook-validate` job that checks out the repository,
installs the latest version of `uv`, and runs `lefthook validate` to
validate Lefthook configurations.