ci: Swap SuperLinter to Full Version (#266)
# Pull Request

## Description

This pull request updates dependencies and configurations to ensure
compatibility and improve code quality checks. The most important
changes include updating the version of the `super-linter` GitHub Action
and increasing the minimum version requirement for `lefthook`.

### Dependency Updates:
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L27-L36):
Updated the `super-linter` GitHub Action from version `v7.3.0` to
`v7.4.0` to incorporate the latest fixes and improvements. Removed the
`EDITORCONFIG_FILE_NAME` environment variable since it is no longer
required.

*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dL2-R2):
Increased the minimum version of `lefthook` from `1.11.12` to `1.11.13`
to ensure compatibility with the latest features.