chore: Set Zizmor persona to auditor (#313)
# Pull Request

## Description

This pull request updates various workflow files to use the latest
version of reusable workflows and modifies the configuration for
`lefthook` and `zizmor` tools. The changes ensure compatibility with
newer versions and streamline the checking process.

### Workflow Updates:
*
[`.github/workflows/clean-caches.yml`](diffhunk://#diff-d0394e4336a74cdfc1d4cff05d056b893ac7ff922eacf4448e104a754f386b8dL15-R15):
Updated to use version `v2025.06.16.01` of the `common-clean-caches.yml`
reusable workflow.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L63-R63):
Updated to use version `v2025.06.16.01` of the `common-code-checks.yml`
and `codeql-analysis.yml` reusable workflows.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L63-R63)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L75-R75)
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfL24-R24):
Updated to use version `v2025.06.16.01` of the
`common-pull-request-tasks.yml` reusable workflow.
*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6L18-R18):
Updated to use version `v2025.06.16.01` of the `common-sync-labels.yml`
reusable workflow.

### Tool Configuration Updates:
*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cL53-R53):
Changed the `zizmor-check` persona from `pedantic` to `auditor`, and
removed the `zizmor-check-sarif` command.
*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dL2-R2):
Updated the `min_version` requirement from `1.11.13` to `1.11.14`.